Back-And-Front/requirements.txt
APScheduler==3.6.3
asgiref==3.7.2
attrs==23.1.0
cachetools==4.2.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coverage==7.2.7
cryptography==41.0.2
defusedxml==0.7.1
discover==0.4.0
distlib==0.3.7
dj-rest-auth==4.0.0
Django==4.0.10
django-cors-headers==3.10.1
django-debug-toolbar==4.0.0
django-dotenv==1.4.2
django-extensions==3.2.1
django-htmx==1.14.0
django-ranged-response==0.2.0
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
django-rest-framework==0.1.0
django-simple-captcha==0.5.17
django-vite==2.1.3
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
filelock==3.12.2
flake8==6.0.0
flake8-isort==6.0.0
greenlet==2.0.2
idna==3.4
inflection==0.5.1
isort==5.12.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
mccabe==0.7.0
mysqlclient==2.2.0
oauthlib==3.2.2
Pillow==10.0.0
pipenv==2023.7.23
platformdirs==3.10.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.8.0
pyTelegramBotAPI==4.7.1
python-telegram-bot==13.14
python3-openid==3.2.0
pytube==15.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
six==1.16.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
telegram==0.0.1
tornado==6.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.2
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.24.2
virtualenv-clone==0.5.7
